Struct supercow::ext::InlineStorage

pub struct InlineStorage<A, B>(_);

Causes the OWNED or SHARED value of a Supercow to be stored inline.

This makes allocation of owned Supercows much faster, at the expense of making the Supercow itself much bigger (since it now must contain the whole object).