Trait supercow::ext::PtrWrite

pub unsafe trait PtrWrite<T: ?Sized>: Copy {
    fn new() -> Self;

    fn store_ptr(&mut self, t: *const T);
}

Optionally stores a pointer to a value.

It is doubtful that there are any types besides () and *mut T which could implement this usefully.

Required methods

fn new() -> Self

Returns an instance of Self with no particular value.

fn store_ptr(&mut self, t: *const T)

Writes the given pointer into self.

Unsafety

The implementation must not inspect the given pointer. This call must not panic.