Trait supercow::ext::SharedFrom

pub unsafe trait SharedFrom<T> {
    fn shared_from(t: T) -> Self;
}

Like std::convert::From, but without the blanket implementations that cause problems for supercow_features!.

Unsafety

The conversion may not invalidate the address returned by T::const_deref() if T is ConstDeref.

Required methods

fn shared_from(t: T) -> Self

Converts the given T to Self.