Type Definition supercow::NonSyncSupercow

type NonSyncSupercow<'a, OWNED, BORROWED = OWNED> = Supercow<'a, OWNED, BORROWED, Box<dyn NonSyncFeatures<'static> + 'static>, BoxedStorage>;

Supercow with the default SHARED changed to NonSyncFeatures, enabling the use of Rc as a shared reference type as well as making it possible to use non-Send or non-Sync BORROWED types easily.

Note that the SHARED type must have 'static lifetime, since this is generally more convenient and makes the Supercow as a whole covariant.

Example

use supercow::{NonSyncSupercow, Supercow};

let x: NonSyncSupercow<u32> = Supercow::owned(42u32);
println!("{}", *x);